JianZhiOffer/E40_最小的k个数_test.go
package offer

import (
	"reflect"
	"testing"
)

func Test_getLeastNumbers(t *testing.T) {
	type args struct {
		arr []int
		k   int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			args: args{
				arr: []int{4, 5, 1, 6, 2, 7, 3, 8},
				k:   4,
			},
			want: []int{1, 2, 3, 4},
		},
		{
			args: args{
				arr: []int{4, 5, 1, 6, 2, 7, 3, 8},
				k:   1,
			},
			want: []int{1},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := getLeastNumbers(tt.args.arr, tt.args.k); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("getLeastNumbers() = %v, want %v", got, tt.want)
			}
		})
	}
}
